initial refactor to mvc
