+++ server/controllers/index.js
@@ -1,1 +1,2 @@
-const models = require ('../models');
+// const models = require ('../models');
+module.exports.products = require('./products.js')
